refactor: mv compilation-config.js config.js into src
