+++ packages/hap-compiler/src/style/process.js
@@ -8,7 +8,7 @@ import fs from 'fs'
 import path from '@jayfate/path'
 import css from 'css'
 import cssWhat from 'css-what'
-import { compileOptionsObject } from '@hap-toolkit/shared-utils/compilation-config'
+import { compileOptionsObject } from '@hap-toolkit/shared-utils/lib/compilation-config'
 import {
   validate as validateDelaration,
   validatePseudoClass,
