+++ packages/hap-dsl-xvm/src/loaders/access-loader.js
@@ -4,7 +4,7 @@
  */
 
 'use strict'
-const globalConfig = require('@hap-toolkit/shared-utils/config')
+const globalConfig = require('@hap-toolkit/shared-utils/lib/config')
 
 /**
  * 添加ViewModel数据校验和处理代码。只有页面组件才需要, 子组件不需要
