+++ packages/hap-dsl-xvm/src/loaders/common/utils.js
@@ -7,7 +7,7 @@ import path from '@jayfate/path'
 import loaderUtils from 'loader-utils'
 import hash from 'hash-sum'
 import { SourceMapGenerator, SourceMapConsumer } from 'source-map'
-import globalConfig from '@hap-toolkit/shared-utils/config'
+import globalConfig from '@hap-toolkit/shared-utils/lib/config'
 
 /**
  * 从文件路径中获取最后一部分（文件名(不包含扩展名)或目录名）
