+++ packages/hap-dsl-xvm/src/loaders/script-loader.js
@@ -6,7 +6,7 @@
 'use strict'
 
 import { parseScript } from '@hap-toolkit/compiler'
-const globalConfig = require('@hap-toolkit/shared-utils/config')
+const globalConfig = require('@hap-toolkit/shared-utils/lib/config')
 
 module.exports = function (source, sourcemap) {
   const { parsed } = parseScript(source)
