+++ packages/hap-dsl-xvm/src/loaders/ux-loader.js
@@ -7,7 +7,7 @@ import path from '@jayfate/path'
 import loaderUtils from 'loader-utils'
 import { parseFragmentsWithCache } from '@hap-toolkit/compiler'
 import validator from '@hap-toolkit/compiler/lib/template/validator'
-import { compileOptionsObject } from '@hap-toolkit/shared-utils/compilation-config'
+import { compileOptionsObject } from '@hap-toolkit/shared-utils/lib/compilation-config'
 
 import {
   processImportFrag,
@@ -18,7 +18,7 @@ import {
 } from './ux-fragment-utils'
 import { getNameByPath, print, ENTRY_TYPE, isUXRender } from './common/utils'
 
-const globalConfig = require('@hap-toolkit/shared-utils/config')
+const globalConfig = require('@hap-toolkit/shared-utils/lib/config')
 
 /**
  * 收集结果
