+++ packages/hap-packager/src/plugins/notify-plugin.js
@@ -7,7 +7,7 @@ import http from 'http'
 import fs from 'fs'
 import { colorconsole, getDeviceInfo } from '@hap-toolkit/shared-utils'
 import { getRecords, getProjectClients } from '@hap-toolkit/shared-utils/lib/record-client'
-import globalConfig from '@hap-toolkit/shared-utils/config'
+import globalConfig from '@hap-toolkit/shared-utils/lib/config'
 
 let clientExists = false
 let isFirstUpdate = true
