+++ packages/hap-packager/src/plugins/resource-plugin.js
@@ -7,9 +7,9 @@ import path from '@jayfate/path'
 import fs from 'fs-extra'
 import aaptjs from '@hap-toolkit/aaptjs'
 import glob from 'glob'
-import eventBus from '@hap-toolkit/shared-utils/event-bus'
+import eventBus from '@hap-toolkit/shared-utils/lib/event-bus'
 import { colorconsole, relateCwd, readJson, logger } from '@hap-toolkit/shared-utils'
-import { compileOptionsObject } from '@hap-toolkit/shared-utils/compilation-config'
+import { compileOptionsObject } from '@hap-toolkit/shared-utils/lib/compilation-config'
 
 import { name } from '../common/info'
 import { updateManifest } from '../common/shared'
