+++ packages/hap-packager/src/plugins/splitchunks-adapt-plugin.js
@@ -6,7 +6,7 @@
 import path from '@jayfate/path'
 import Compilation from 'webpack/lib/Compilation'
 import { getEntryFiles } from '../common/info'
-import { compileOptionsMeta } from '@hap-toolkit/shared-utils/compilation-config'
+import { compileOptionsMeta } from '@hap-toolkit/shared-utils/lib/compilation-config'
 import { isEmptyObject } from '../common/utils'
 
 let ConcatSource
