+++ packages/hap-packager/src/plugins/zip-plugin.js
@@ -6,13 +6,13 @@
 import fs from 'fs-extra'
 import path from '@jayfate/path'
 import { colorconsole, readJson } from '@hap-toolkit/shared-utils'
-import globalConfig from '@hap-toolkit/shared-utils/config'
+import globalConfig from '@hap-toolkit/shared-utils/lib/config'
 import { sortFilesBy, lsdirdeep, genPriorities } from '../common/utils'
 import { getSkeletonConfig } from '../common/info'
 import {
   compileOptionsMeta,
   compileOptionsObject
-} from '@hap-toolkit/shared-utils/compilation-config'
+} from '@hap-toolkit/shared-utils/lib/compilation-config'
 import { DIGEST_ZIP_PATH } from '../common/constant'
 import {
   createPackagesDefinition,
