+++ packages/hap-server/src/server.js
@@ -9,7 +9,7 @@ import opn from 'opn'
 import portfinder from 'portfinder'
 import { outputQRCodeOnTerminal, getIPv4IPAddress, colorconsole } from '@hap-toolkit/shared-utils'
 import { clearProjectRecord } from '@hap-toolkit/shared-utils/lib/record-client'
-import globalConfig from '@hap-toolkit/shared-utils/config'
+import globalConfig from '@hap-toolkit/shared-utils/lib/config'
 
 let server = null
 export async function launch(conf, moduler) {
