+++ packages/hap-shared-utils/src/buildMode/BuildModeManager.js
@@ -7,7 +7,7 @@ import fs from 'fs'
 import path from '@jayfate/path'
 import { readJson } from '../index'
 import { initProjectConfig, CONFIG_FILE } from './util'
-import eventBus from '../../event-bus'
+import eventBus from '../event-bus'
 
 const { PACKAGER_BUILD_DONE } = eventBus
 /**
