+++ packages/hap-shared-utils/src/compilation-config.js
