+++ packages/hap-shared-utils/src/config.js
