+++ packages/hap-shared-utils/src/event-bus.js
