--- packages/hap-toolkit/config.js
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2021-present, the hapjs-platform Project Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-module.exports = {
-  projectPath: process.cwd()
-}
