+++ packages/hap-toolkit/gen-webpack-conf/helpers.js
@@ -6,7 +6,7 @@
 const fs = require('fs-extra')
 const path = require('@jayfate/path')
 const builtinList = require('module').builtinModules
-const globalConfig = require('@hap-toolkit/shared-utils/config')
+const globalConfig = require('@hap-toolkit/shared-utils/lib/config')
 
 const { readJson, colorconsole } = require('@hap-toolkit/shared-utils')
